Added drag and drop

+++ controllers/tasksController.js
@@ -1,5 +1,6 @@
 const { controllerWrapper } = require('../decorators');
 const Task = require('../models/task');
+const Column = require('../models/column');
 
 const { HttpError } = require('../helpers');
 
@@ -27,8 +28,19 @@ const getAllTasks = async (req, res) => {
 };
 
 const addTask = async (req, res) => {
+  // ===-VR-===
+  const { parentColumn } = req.body;
+  // ===-VR-===
+
   const result = await Task.create({ ...req.body });
 
+  // ===-VR-===
+  await Column.findByIdAndUpdate(
+    parentColumn,
+    { $push: { taskOrder: result._id } },
+    { safe: true, upsert: true, new: true }
+  );
+  // ===-VR-===
   res.status(201).json(result);
 };
 
@@ -44,6 +56,18 @@ const updateTask = async (req, res) => {
 const deleteTask = async (req, res) => {
   const { id } = req.params;
 
+  // ===-VR-===
+  const task = await Task.findById(id);
+
+  if (!task) {
+    throw HttpError(404, 'Task not found');
+  }
+  const qwe = await Column.findByIdAndUpdate(task.parentColumn, {
+    $pull: { taskOrder: task._id },
+  });
+  console.log(qwe);
+  // ===-VR-===
+
   const result = await Task.findByIdAndDelete(id);
   if (!result) {
     throw HttpError(404, 'Task not found');
@@ -51,9 +75,30 @@ const deleteTask = async (req, res) => {
   res.status(204).json({ message: `Task with id: ${id} deleted` });
 };
 
+// ===-VR-===
+const moveTask = async (req, res) => {
+  const { id } = req.params;
+  const { columnSourceOrder, columnDestinationOrder } = req.body;
+  const destinitionColumnId = Object.keys(columnDestinationOrder)[0];
+  const sourceColumnId = Object.keys(columnSourceOrder)[0];
+  const destinationColumnArr = Object.values(columnDestinationOrder)[0];
+  const sourceColumnArr = Object.values(columnSourceOrder)[0];
+
+  await Column.findByIdAndUpdate(sourceColumnId, { taskOrder: sourceColumnArr });
+  await Column.findByIdAndUpdate(destinitionColumnId, { taskOrder: destinationColumnArr });
+
+  const task = await Task.findByIdAndUpdate(id, { parentColumn: destinitionColumnId }, { new: true });
+
+  res.status(201).json(task);
+};
+// ===-VR-===
+
 module.exports = {
   getAllTasks: controllerWrapper(getAllTasks),
   addTask: controllerWrapper(addTask),
   updateTask: controllerWrapper(updateTask),
   deleteTask: controllerWrapper(deleteTask),
+  // ===-VR-===
+  moveTask: controllerWrapper(moveTask),
+  // ===-VR-===
 };
